move testing utilities to `mod test`
